backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
httpx>=0.27.0
openai>=1.55.0
numpy>=1.26.0
pandas>=2.2.0
langgraph>=0.2.0
faiss-cpu>=1.8.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
